requirements.txt
Pillow==2.9.0
PyPDF2==1.25.1
Wand==0.4.1
zbar
